chore: Refactor layout and styles for better organization
In this commit, style rules of the layout files were moved to their respective external CSS files to maintain separation of concerns. This also included adjustments to padding values and background properties for better appearance. Furthermore, redundant code in the home.njk layout file was cleaned up, and classes were renamed for clarity and consistency.