chore: Improve mobile accessibility.
Adjust video embedding width and update CSS for code display

This commit modifies the layout of the video in the blog post, changing the width to 100% to enhance responsiveness. The CSS style file is also updated to set the maximum width for the non-language specific code and language-specific code elements, improving the appearance and readability of code blocks in the post.