fix: show external link icon
The commit includes the adjustment of CSS rules for better web page presentation. This primarily deals with the enhancement of the blog's video embedding width for increased responsiveness, along with improvement on the maximum width set for non-language specific and language specific code elements, seeking to augment the readability of the code blocks displayed in the post.