feat: Add theme attribute to IT articles
This update introduces a new attribute, "theme", for each of the IT slang article entries under 'gpt.json'. This will allow easier categorization and grouping of articles, making it simpler to navigate through the content. Please note that the value is "funny jargon" for each entry as of this update.