feat: Add blog image on sidebar area if needed.
This update adds functionality to display an image in the table of contents section of blog posts. Additional properties (path, imageEnabled, imageUrl) were included in the Toc function, and conditional rendering was applied for the image. Changes also involve the render method and the modifications in the post template file to pass the necessary values.