Add Assistants API with Stream blog (#1159)
* feat: Add new blog post about improving UX with OpenAI Assistant API

This commit includes a new blog post about using OpenAI's Assistant API Stream responses to build real-time experiences, with detailed instructions on how to create the conversation script in Node.js. It also includes an update to the tags in a previous blog post.

* chore: Add image attribute to OpenAI Assistant API blog post

This update adds the 'image' attribute to the metadata section of the OpenAI Assistant API blog post. This important detail will ensure that the blog post correctly displays an image when viewed on the website.

* chore: update doc

* chore: update doc

* chore: update doc