feat: Add new blog post about using AWS CodeBuild with GitHub Actions
This commit includes a new blog post explaining how to use AWS CodeBuild as a self hosted runner for GitHub Actions. There was also an update made to an old blog post with a new link to AWS CodeBuild's official documentation. Lastly, the list of blog posts on the CI/CD page has been updated with a link to the new blog post.