feat: Add new blog post about running Playwright on Lambda
This commit adds a new blog post detailing how to use Playwright, a popular tool for browser automation, on AWS Lambda. The blog post contains step-by-step instructions as well as code snippets to guide the reader on how to set it up. It describes two methods: using a Lambda layer and using a Lambda container.