Add ranking.json and Lume Update (#1645)
* Update lume dependency to version 2.4.3

Upgraded the lume dependency from version 2.2.4 to 2.4.3 in deno.json. This ensures access to the latest features, bug fixes, and improvements provided by the newer version.

* Update Node.js version in Netlify config to 22

Upgraded the `NODE_VERSION` from 18 to 22 in the Netlify configuration file. This ensures compatibility with the latest Node.js features and improvements.

* Add ranking JSON generator for PVRanking data

Introduces a generator that creates a `ranking.json` feed from PVRanking data. The output includes title and URL mappings, formatted as JSON, and is excluded from default generation.